+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KRED5HWZCQCY0MT9N09P1V7K"
+        "01KREDMNRRN613Q55H4AN6BDWS"
     )
     r"""The unique identifier of the data source"""